Ticket #157 solved. Alle Id's zijn nu Long. Overflows van de Long bij een insert of update in de database zouden een SQL-error moeten geven. Indien dit gebeurt is de database index domein 'vol'.
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWO-commons-lib/branches/working@10753 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/commons/persistence/entities/PersistentApplet.java
@@ -38,7 +38,7 @@ public class PersistentApplet implements PersistentUpdate<PersistentApplet>, Ser
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "appletID", nullable = false)
-    private Integer appletID;
+    private Long appletID;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 128)
@@ -70,22 +70,22 @@ public class PersistentApplet implements PersistentUpdate<PersistentApplet>, Ser
     public PersistentApplet() {
     }
 
-    public PersistentApplet(Integer appletID) {
+    public PersistentApplet(Long appletID) {
         this.appletID = appletID;
     }
 
-    public PersistentApplet(Integer appletID, String appletName, String classname, String jarname) {
+    public PersistentApplet(Long appletID, String appletName, String classname, String jarname) {
         this.appletID = appletID;
         this.appletName = appletName;
         this.classname = classname;
         this.jarname = jarname;
     }
 
-    public Integer getAppletID() {
+    public Long getAppletID() {
         return appletID;
     }
 
-    public void setAppletID(Integer appletID) {
+    public void setAppletID(Long appletID) {
         this.appletID = appletID;
     }
